+++ R/data_rename.R
@@ -111,6 +111,17 @@ data_rename <- function(data,
     verbose = verbose
   )
 
+  # Forbid partially named "select",
+  # Ex: if select = c("foo" = "Species", "Sepal.Length") then the 2nd name and
+  # 2nd value are "Sepal.Length"
+  if (!is.null(names(select))) {
+    for (i in seq_along(select)) {
+      if (names(select)[i] == select[i]) {
+        insight::format_error("When `select` is a named vector, all elements must be named.")
+      }
+    }
+  }
+
   # check if `select` has names, and if so, use as "replacement"
   if (!is.null(names(select))) {
     replacement <- names(select)